requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-bootstrap4==3.0.1
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-crispy-forms==1.11.2
django-des==2.4.1
django-embed-video==1.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jazzmin==2.4.7
django-js-asset==1.2.2
django-solo==1.1.3
easy-thumbnails==2.7.1
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
